requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pytesseract==0.3.4
soupsieve==2.0.1
Werkzeug==1.0.1
